+++ web-console/src/lib/compositions/streaming/import/useDefaultRows.ts
@@ -1,6 +1,6 @@
 // Generates rows and inserts them into a table.
 
-import { getValueParser, Row } from '$lib/functions/ddl'
+import { Row } from '$lib/functions/ddl'
 import { ColumnType, Field, Relation } from '$lib/services/manager'
 import dayjs from 'dayjs'
 import { Dispatch, MutableRefObject, SetStateAction, useCallback } from 'react'
@@ -12,7 +12,7 @@ export const getDefaultValue = (columntype: ColumnType) =>
   match(columntype)
     .with({ type: 'BOOLEAN' }, () => false)
     .with({ type: 'TINYINT' }, { type: 'SMALLINT' }, { type: 'INTEGER' }, { type: 'BIGINT' }, () => 0)
-    .with({ type: 'DECIMAL' }, { type: 'FLOAT' }, { type: 'DOUBLE' }, () => 0.0)
+    .with({ type: 'DECIMAL' }, { type: 'NUMERIC' }, { type: 'FLOAT' }, { type: 'DOUBLE' }, () => 0.0)
     .with({ type: 'VARCHAR' }, { type: 'CHAR' }, () => '')
     .with({ type: 'TIME' }, () => dayjs(new Date()))
     .with({ type: 'TIMESTAMP' }, () => dayjs(new Date()))
@@ -35,8 +35,7 @@ function useDefaultRows(
         for (let i = 0; i < rowCount; i++) {
           const row: Row = { genId: curRowCount + i, weight: 1, record: {} }
           relation.fields.forEach((field: Field) => {
-            const valueParser = getValueParser(field.columntype)
-            row.record[field.name] = valueParser(getDefaultValue(field.columntype))
+            row.record[field.name] = getDefaultValue(field.columntype)
           })
           newRows.push(row)
         }